requirements.txt
pandas
dotenv
opencv-contrib-python
pillow
scikit-image
scikit-learn
torch
transformers
imagehash
av
langchain
langchain_community
langchain_google_genai
langgraph